+++ cognite/neat/core/rules/models.py
@@ -7,7 +7,6 @@ from datetime import datetime
 from pathlib import Path
 from typing import ClassVar, Dict, Optional
 
-from graphql import GraphQLBoolean, GraphQLFloat, GraphQLInt, GraphQLString
 from pydantic import (
     BaseModel,
     ConfigDict,
@@ -33,17 +32,17 @@ __all__ = ["Class", "Instance", "Metadata", "Prefixes", "Property", "Resource", 
 
 # mapping of XSD types to Python and GraphQL types
 DATA_TYPE_MAPPING = {
-    "boolean": {"python": "bool", "GraphQL": GraphQLBoolean},
-    "float": {"python": "float", "GraphQL": GraphQLFloat},
-    "integer": {"python": "int", "GraphQL": GraphQLInt},
-    "nonPositiveInteger": {"python": "int", "GraphQL": GraphQLInt},
-    "nonNegativeInteger": {"python": "int", "GraphQL": GraphQLInt},
-    "negativeInteger": {"python": "int", "GraphQL": GraphQLInt},
-    "long": {"python": "int", "GraphQL": GraphQLInt},
-    "string": {"python": "str", "GraphQL": GraphQLString},
-    "anyURI": {"python": "str", "GraphQL": GraphQLString},
-    "normalizedString": {"python": "str", "GraphQL": GraphQLString},
-    "token": {"python": "str", "GraphQL": GraphQLString},
+    "boolean": {"python": "bool", "GraphQL": "Boolean"},
+    "float": {"python": "float", "GraphQL": "Float"},
+    "integer": {"python": "int", "GraphQL": "Int"},
+    "nonPositiveInteger": {"python": "int", "GraphQL": "Int"},
+    "nonNegativeInteger": {"python": "int", "GraphQL": "Int"},
+    "negativeInteger": {"python": "int", "GraphQL": "Int"},
+    "long": {"python": "int", "GraphQL": "Int"},
+    "string": {"python": "str", "GraphQL": "String"},
+    "anyURI": {"python": "str", "GraphQL": "String"},
+    "normalizedString": {"python": "str", "GraphQL": "String"},
+    "token": {"python": "str", "GraphQL": "String"},
     # Graphql does not have a datetime type this is CDF specific
     "dateTime": {"python": "datetime", "GraphQL": "Timestamp"},
 }